(feat) Add config and weights loading. Change RMSNorm weight from fp16 to fp32.

+++ include/kernels/rmsnorm.cuh
@@ -7,16 +7,16 @@
 // rms(x) = sqrt(mean(x^2) + eps)
 //
 // Shape: x, out -> [rows, cols]  (half)
-//        weight -> [cols]        (half)
+//        weight -> [cols]        (float) — stored in FP32 to avoid BF16->FP16 range loss
 // FP16 I/O, FP32 accumulation internally.
 //
 // Supported configs (Qwen3):
 //   cols=128  -> head_dim (q_norm / k_norm)  -> 1 warp  (32 threads)
 //   cols=1024 -> hidden_size (layernorm)     -> 4 warps (128 threads)
 void launch_rmsnorm(
-    half*        out,
-    const half*  x,
-    const half*  weight,
+    half*         out,
+    const half*   x,
+    const float*  weight,
     int          rows,
     int          cols,
     float        eps,
